@@ -8,24 +8,23 @@
         "name" : "Alvaro Videla"
     }],
     "require": {
-        "php":                          "^7.1|^8.0",
+        "php":                          "^7.4|^8.0",
 
-        "symfony/dependency-injection": "^4.3|^5.0",
-        "symfony/event-dispatcher":     "^4.3|^5.0",
-        "symfony/config":               "^4.3|^5.0",
-        "symfony/yaml":                 "^4.3|^5.0",
-        "symfony/console":              "^4.3|^5.0",
+        "symfony/dependency-injection": "^4.4|^5.3|^6.0",
+        "symfony/event-dispatcher":     "^4.4|^5.3|^6.0",
+        "symfony/config":               "^4.4|^5.3|^6.0",
+        "symfony/yaml":                 "^4.4|^5.3|^6.0",
+        "symfony/console":              "^4.4|^5.3|^6.0",
         "php-amqplib/php-amqplib":      "^2.12.2|^3.0",
-        "psr/log":                      "^1.0",
-        "symfony/http-kernel":          "^4.4|^5.0",
-        "symfony/framework-bundle":     "^4.4|^5.0",
-        "swaggest/json-schema":         "^0.12.38"
+        "psr/log":                      "^1.0 || ^2.0 || ^3.0",
+        "symfony/http-kernel":          "^4.4|^5.3|^6.0",
+        "symfony/framework-bundle":     "^4.4|^5.3|^6.0"
     },
     "require-dev": {
-        "symfony/serializer":           "^4.3|^5.0",
-        "phpunit/phpunit":              "^7.5|^8.5",
-        "phpstan/phpstan":              "^0.12.11",
-        "phpstan/phpstan-phpunit":      "^0.12.6"
+        "symfony/serializer":           "^4.4|^5.3|^6.0",
+        "phpunit/phpunit":              "^9.5",
+        "phpstan/phpstan":              "^1.2",
+        "phpstan/phpstan-phpunit":      "^1.0"
     },
     "replace": {
         "oldsound/rabbitmq-bundle": "self.version",
