@@ -4,21 +4,21 @@ namespace OldSound\RabbitMqBundle\RabbitMq;
 
 class AmqpPartsHolder
 {
-	protected $parts;
+    protected $parts;
 
-	public function __construct()
-	{
-		$this->parts = array();
-	}
+    public function __construct()
+    {
+        $this->parts = [];
+    }
 
-	public function addPart($type, BaseAmqp $part)
-	{
-		$this->parts[$type][] = $part;
-	}
+    public function addPart($type, BaseAmqp $part)
+    {
+        $this->parts[$type][] = $part;
+    }
 
-	public function getParts($type)
-	{
+    public function getParts($type)
+    {
         $type = (string) $type;
-		return isset($this->parts[$type]) ? $this->parts[$type] : array();
-	}
+        return $this->parts[$type] ?? [];
+    }
 }
