@@ -46,7 +46,7 @@ class Producer extends BaseAmqp implements ProducerInterface
 
     protected function getBasicProperties()
     {
-        return array('content_type' => $this->contentType, 'delivery_mode' => $this->deliveryMode);
+        return ['content_type' => $this->contentType, 'delivery_mode' => $this->deliveryMode];
     }
 
     public function validateMessage($msg)
@@ -69,7 +69,7 @@ class Producer extends BaseAmqp implements ProducerInterface
      * @param array $additionalProperties
      * @param array $headers
      */
-    public function publish($msgBody, $routingKey = null, $additionalProperties = array(), array $headers = null)
+    public function publish($msgBody, $routingKey = null, $additionalProperties = [], array $headers = null)
     {
         if ($this->validator != null){
             $this->validateMessage($msgBody);
@@ -88,13 +88,13 @@ class Producer extends BaseAmqp implements ProducerInterface
 
         $real_routingKey = $routingKey !== null ? $routingKey : $this->defaultRoutingKey;
         $this->getChannel()->basic_publish($msg, $this->exchangeOptions['name'], (string)$real_routingKey);
-        $this->logger->debug('AMQP message published', array(
-            'amqp' => array(
+        $this->logger->debug('AMQP message published', [
+            'amqp' => [
                 'body' => $msgBody,
                 'routingkeys' => $routingKey,
                 'properties' => $additionalProperties,
-                'headers' => $headers
-            )
-        ));
+                'headers' => $headers,
+            ],
+        ]);
     }
 }
