@@ -27,7 +27,8 @@ abstract class BaseConsumerCommand extends BaseRabbitMqCommand
             // Halt consumer if waiting for a new message from the queue
             try {
                 $this->consumer->stopConsuming();
-            } catch (AMQPTimeoutException $e) {}
+            } catch (AMQPTimeoutException $e) {
+            }
         }
     }
 
@@ -80,9 +81,9 @@ abstract class BaseConsumerCommand extends BaseRabbitMqCommand
                 throw new \BadFunctionCallException("Function 'pcntl_signal' is referenced in the php.ini 'disable_functions' and can't be called.");
             }
 
-            pcntl_signal(SIGTERM, array(&$this, 'stopConsumer'));
-            pcntl_signal(SIGINT, array(&$this, 'stopConsumer'));
-            pcntl_signal(SIGHUP, array(&$this, 'restartConsumer'));
+            pcntl_signal(SIGTERM, [&$this, 'stopConsumer']);
+            pcntl_signal(SIGINT, [&$this, 'stopConsumer']);
+            pcntl_signal(SIGHUP, [&$this, 'restartConsumer']);
         }
 
         if (defined('AMQP_DEBUG') === false) {
