@@ -12,7 +12,7 @@ use PhpAmqpLib\Message\AMQPMessage;
  */
 class BeforeProcessingMessageEvent extends AMQPEvent
 {
-    const NAME = AMQPEvent::BEFORE_PROCESSING_MESSAGE;
+    public const NAME = AMQPEvent::BEFORE_PROCESSING_MESSAGE;
 
     /**
      * BeforeProcessingMessageEvent constructor.
