@@ -11,7 +11,7 @@ use OldSound\RabbitMqBundle\RabbitMq\Consumer;
  */
 class OnConsumeEvent extends AMQPEvent
 {
-    const NAME = AMQPEvent::ON_CONSUME;
+    public const NAME = AMQPEvent::ON_CONSUME;
 
     /**
      * OnConsumeEvent constructor.
