@@ -11,16 +11,16 @@ interface QueueOptionsProviderInterface
 {
     /**
      * Return queue options
-     * 
+     *
      * Example:
      * array(
      *   'name' => 'example_context',
      *   'durable' => true,
      *   'routing_keys' => array('key.*')
      * )
-     * 
+     *
      * @return array
-     * 
+     *
      */
     public function getQueueOptions($context = null);
 }
