@@ -141,7 +141,7 @@ class Binding extends BaseAmqp
         $method  = ($this->destinationIsExchange) ? 'exchange_bind' : 'queue_bind';
         $channel = $this->getChannel();
         call_user_func(
-            array($channel, $method),
+            [$channel, $method],
             $this->destination,
             $this->exchange,
             $this->routingKey,
