@@ -9,27 +9,27 @@ interface ConsumerInterface
     /**
      * Flag for message ack
      */
-    const MSG_ACK = 1;
+    public const MSG_ACK = 1;
 
     /**
      * Flag single for message nack and requeue
      */
-    const MSG_SINGLE_NACK_REQUEUE = 2;
+    public const MSG_SINGLE_NACK_REQUEUE = 2;
 
     /**
      * Flag for reject and requeue
      */
-    const MSG_REJECT_REQUEUE = 0;
+    public const MSG_REJECT_REQUEUE = 0;
 
     /**
      * Flag for reject and drop
      */
-    const MSG_REJECT = -1;
+    public const MSG_REJECT = -1;
 
     /**
      * Flag for consumers that wants to handle ACKs on their own
      */
-    const MSG_ACK_SENT = -2;
+    public const MSG_ACK_SENT = -2;
 
     /**
      * @param AMQPMessage $msg The message
