@@ -4,4 +4,4 @@ namespace OldSound\RabbitMqBundle\RabbitMq\Exception;
 
 class QueueNotFoundException extends \RuntimeException
 {
-} 
+}
