@@ -12,28 +12,28 @@ class RpcServerTest extends TestCase
     {
         /** @var RpcServer $server */
         $server = $this->getMockBuilder('\OldSound\RabbitMqBundle\RabbitMq\RpcServer')
-            ->setMethods(array('sendReply', 'maybeStopConsumer'))
+            ->setMethods(['sendReply', 'maybeStopConsumer'])
             ->disableOriginalConstructor()
             ->getMock();
         $message = $this->getMockBuilder('\PhpAmqpLib\Message\AMQPMessage')
-            ->setMethods( array('get'))
+            ->setMethods(['get'])
             ->getMock();
         $message->setChannel(
             $this->getMockBuilder('\PhpAmqpLib\Channel\AMQPChannel')
-                ->setMethods(array())->setConstructorArgs(array())
+                ->setMethods([])->setConstructorArgs([])
                 ->setMockClassName('')
                 ->disableOriginalConstructor()
                 ->getMock()
         );
         $message->setDeliveryTag(0);
-        $server->setCallback(function() {
+        $server->setCallback(function () {
             return 'message';
         });
         $serializer = $this->getMockBuilder('\Symfony\Component\Serializer\SerializerInterface')
-            ->setMethods(array('serialize', 'deserialize'))
+            ->setMethods(['serialize', 'deserialize'])
             ->getMock();
         $serializer->expects($this->once())->method('serialize')->with('message', 'json');
-        $server->setSerializer(function($data) use ($serializer) {
+        $server->setSerializer(function ($data) use ($serializer) {
             $serializer->serialize($data, 'json');
         });
         $server->processMessage($message);
